requirements.txt
whitenoise==3.3.1
Django==2.0
dj_database_url==0.4.2
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.3.1
django-filter==1.0.4
coreapi==2.3.3
django-import-export==0.6.1
